The webview delegate is dead. Long live the webview client
