+++ sdk/forms/src/main/java/com/klaviyo/forms/KlaviyoWebViewClient.kt
@@ -11,12 +11,7 @@ import android.webkit.WebResourceResponse
 import android.webkit.WebView
 import android.webkit.WebViewClient
 import androidx.core.content.ContextCompat.startActivity
-import androidx.webkit.JavaScriptReplyProxy
-import androidx.webkit.WebMessageCompat
-import androidx.webkit.WebViewCompat
-import androidx.webkit.WebViewFeature.WEB_MESSAGE_LISTENER
-import com.klaviyo.analytics.Klaviyo
-import com.klaviyo.analytics.networking.ApiClient
+import androidx.core.net.toUri
 import com.klaviyo.core.Registry
 import com.klaviyo.core.config.Clock
 import com.klaviyo.core.lifecycle.ActivityEvent
@@ -27,13 +22,13 @@ import java.io.BufferedReader
 /**
  * Manages [KlaviyoWebView] to power in-app forms
  */
-internal class KlaviyoWebViewDelegate : WebViewClient(), WebViewCompat.WebMessageListener {
+internal class KlaviyoWebViewClient : WebViewClient() {
+    val nativeBridge: BridgeMessageHandler = BridgeMessageHandler(this)
+
     /**
      * Defines origin(s) for which this delegate should be used
      */
     val allowedOrigin: Set<String> get() = setOf(Registry.config.baseUrl)
-
-    val bridgeName = "KlaviyoNativeBridge"
 
     private val activity: Activity? get() = Registry.lifecycleMonitor.currentActivity
 
@@ -56,6 +51,8 @@ internal class KlaviyoWebViewDelegate : WebViewClient(), WebViewCompat.WebMessag
         /**
          * This closes the form on rotation, which we can detect with the local field
          * We wait for a change, see if it's different from the current, and close an open webview
+         *
+         * TODO move to presentation manager
          */
         Registry.lifecycleMonitor.onActivityEvent {
             if (it is ActivityEvent.ConfigurationChanged) {
@@ -76,7 +73,7 @@ internal class KlaviyoWebViewDelegate : WebViewClient(), WebViewCompat.WebMessag
     fun initializeWebView() = webView ?: KlaviyoWebView().also { webView ->
         this.webView = webView
 
-        val klaviyoJsUrl = Uri.parse(Registry.config.baseCdnUrl)
+        val klaviyoJsUrl = Registry.config.baseCdnUrl.toUri()
             .buildUpon()
             .path("onsite/js/klaviyo.js")
             .appendQueryParameter("company_id", Registry.config.apiKey)
@@ -90,7 +87,7 @@ internal class KlaviyoWebViewDelegate : WebViewClient(), WebViewCompat.WebMessag
             .use(BufferedReader::readText)
             .replace("SDK_NAME", Registry.config.sdkName)
             .replace("SDK_VERSION", Registry.config.sdkVersion)
-            .replace("BRIDGE_NAME", this.bridgeName)
+            .replace("BRIDGE_NAME", nativeBridge.name)
             .replace("BRIDGE_HANDSHAKE", handShakeData)
             .replace("KLAVIYO_JS_URL", klaviyoJsUrl.toString())
             .replace("FORMS_ENVIRONMENT", Registry.config.formEnvironment.templateName)
@@ -151,55 +148,22 @@ internal class KlaviyoWebViewDelegate : WebViewClient(), WebViewCompat.WebMessag
                 data = request.url
                 addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
             }
-            activity?.let {
-                startActivity(it, intent, null)
-            } ?: run {
-                Registry.log.warning("Unable to launch external browser - null activity reference")
-            }
+            Registry.config.applicationContext.startActivity(intent, null)
             return true
         }
         return super.shouldOverrideUrlLoading(view, request)
     }
 
     /**
-     * When [WEB_MESSAGE_LISTENER] is supported, messages sent over the Native Bridge from JS are received here
-     */
-    override fun onPostMessage(
-        view: WebView,
-        message: WebMessageCompat,
-        sourceOrigin: Uri,
-        isMainFrame: Boolean,
-        replyProxy: JavaScriptReplyProxy
-    ) = message.data?.let { postMessage(it) } ?: Unit
-
-    /**
-     * When [WEB_MESSAGE_LISTENER] is NOT supported, messages sent over the Native Bridge from JS are received here
-     */
-    @android.webkit.JavascriptInterface
-    fun postMessage(message: String) = try {
-        Registry.log.debug("JS interface postMessage $message")
-        when (val bridgeMessage = BridgeMessage.decodeWebviewMessage(message)) {
-            BridgeMessage.HandShook -> handShook()
-            is BridgeMessage.Show -> show()
-            is BridgeMessage.AggregateEventTracked -> createAggregateEvent(bridgeMessage)
-            is BridgeMessage.ProfileEvent -> createProfileEvent(bridgeMessage)
-            is BridgeMessage.DeepLink -> deepLink(bridgeMessage)
-            is BridgeMessage.Close -> close()
-            is BridgeMessage.Abort -> abort(bridgeMessage.reason)
-        }
-    } catch (e: Exception) {
-        Registry.log.warning("Failed to relay webview message: $message", e)
-    }.let { }
-
-    /**
      * Handle a [BridgeMessage.HandShook] message by clearing the timer
      */
-    private fun handShook() = handshakeTimer?.cancel()
+    fun onJsHandshakeCompleted() = handshakeTimer?.cancel()
 
     /**
-     * Handle a [BridgeMessage.Show] message by displaying the webview before the form animates in
+     * Show the webview in the current activity
+     * TODO move to presentation manager
      */
-    private fun show() = webView?.let { webView ->
+    fun show() = webView?.let { webView ->
         activity?.window?.decorView?.post { decorView ->
             decorView.findViewById<ViewGroup>(android.R.id.content).addView(webView)
             webView.visibility = View.VISIBLE
@@ -211,38 +175,10 @@ internal class KlaviyoWebViewDelegate : WebViewClient(), WebViewCompat.WebMessag
     }
 
     /**
-     * Handle a [BridgeMessage.AggregateEventTracked] message by creating an API call
+     * Remove the webview from the current activity and destroy it
+     * TODO move to presentation manager
      */
-    private fun createAggregateEvent(message: BridgeMessage.AggregateEventTracked) =
-        Registry.get<ApiClient>().enqueueAggregateEvent(message.payload)
-
-    /**
-     * Handle a [BridgeMessage.ProfileEvent] message by creating an API call
-     */
-    private fun createProfileEvent(message: BridgeMessage.ProfileEvent) =
-        Klaviyo.createEvent(message.event)
-
-    /**
-     * Handle a [BridgeMessage.DeepLink] message by broadcasting an intent to the host app
-     * similar to how we handle deep links from a notification
-     */
-    private fun deepLink(messageType: BridgeMessage.DeepLink) = activity?.let { activity ->
-        activity.startActivity(
-            Intent().apply {
-                data = Uri.parse(messageType.route)
-                action = Intent.ACTION_VIEW
-                setPackage(activity.packageName)
-                addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP)
-            }
-        )
-    } ?: Registry.log.warning(
-        "Failed to launch deeplink ${messageType.route} - null activity reference"
-    )
-
-    /**
-     * Handle a [BridgeMessage.Close] message by detaching and destroying the [KlaviyoWebView]
-     */
-    private fun close() = webView?.let { webView ->
+    fun close() = webView?.let { webView ->
         handshakeTimer?.cancel()
         activity?.window?.decorView?.post {
             Registry.log.verbose("Clear IAF WebView reference")
@@ -255,13 +191,6 @@ internal class KlaviyoWebViewDelegate : WebViewClient(), WebViewCompat.WebMessag
         }
     } ?: run {
         Registry.log.warning("Unable to close IAF - null WebView reference")
-    }
-
-    /**
-     * Handle a [BridgeMessage.Abort] message by logging the reason and destroying the webview
-     */
-    private fun abort(reason: String) = close().also {
-        Registry.log.info("IAF aborted, reason: $reason")
     }
 
     /**
